fix: replaceColor on SVG images, refactor previous fix, fixes #1519
